Refactoring seperated subscriber and consumer into seperate module and HMI code is now put in lidar-common module
