+++ hmi-consumer/src/main/java/eu/arrowhead/lidar/consumer/HMIConsumer.java
@@ -12,6 +12,8 @@ import eu.arrowhead.common.model.OrchestrationFlags;
 import eu.arrowhead.common.model.ServiceMetadata;
 import eu.arrowhead.common.model.ServiceRequestForm;
 import eu.arrowhead.lidar.common.LidarReadout;
+import eu.arrowhead.lidar.common.hmi.HMIWindow;
+import eu.arrowhead.lidar.common.hmi.LidarView;
 
 import javax.ws.rs.core.Response;
 import java.util.Timer;
@@ -23,13 +25,12 @@ import java.util.TimerTask;
  */
 public class HMIConsumer extends ArrowheadApplication {
 
-    public static HMIWindow window;
+    public static void main(String[] args) throws ArrowheadException {
+        new HMIConsumer(args).start();
+    }
 
-    public HMIConsumer(String[] args) throws ArrowheadException {
+    private HMIConsumer(String[] args) throws ArrowheadException {
         super(args);
-
-        window = new HMIWindow();
-        window.setVisible(true);
     }
 
     @Override
@@ -39,11 +40,11 @@ public class HMIConsumer extends ArrowheadApplication {
 
         final ArrowheadSystem me = ArrowheadSystem.createFromProperties();
         final ServiceRequestForm srf = new ServiceRequestForm.Builder(me)
-                .requestedService("lidargen", "json", getProps().isSecure())
+                .requestedService("lidargen", "JSON", getProps().isSecure())
                 .metadata(ServiceMetadata.Keys.UNIT, "meters")
                 .flag(OrchestrationFlags.Flags.OVERRIDE_STORE, true)
                 .flag(OrchestrationFlags.Flags.PING_PROVIDERS, false)
-                .flag(OrchestrationFlags.Flags.METADATA_SEARCH, true)
+                .flag(OrchestrationFlags.Flags.METADATA_SEARCH, false)
                 .flag(OrchestrationFlags.Flags.ENABLE_INTER_CLOUD, true)
                 .build();
         log.info("Service Request payload: " + ArrowheadConverter.json().toString(srf));
@@ -69,9 +70,5 @@ public class HMIConsumer extends ArrowheadApplication {
     @Override
     protected void onStop() {
 
-    }
-
-    public static void main(String[] args) throws ArrowheadException {
-        new HMIConsumer(args).start();
     }
 }
