+++ hmi-subscriber/default.conf
@@ -7,7 +7,7 @@ cert_dir=config/certificates
 secure=false
 
 # System registry entry parameters
-system_name=hmi-consumer
+system_name=hmi-subscriber
 
 # Key and trust stores
 bootstrap=true
@@ -16,3 +16,8 @@ keystorepass=123456
 truststore=truststore.p12
 truststorepass=123456
 keypass=123456
+
+# Subscriber parameters
+feedback=true
+feedback_uri=publisher/feedback
+event_type=lidargen
