+++ hmi-subscriber/src/main/java/eu/arrowhead/lidar/subscriber/HMISubscriberResource.java
@@ -1,4 +1,4 @@
-package eu.arrowhead.lidar.consumer;
+package eu.arrowhead.lidar.subscriber;
 
 import eu.arrowhead.common.api.ArrowheadConverter;
 import eu.arrowhead.common.api.server.ArrowheadHttpServer;
@@ -29,7 +29,6 @@ public class HMISubscriberResource extends ArrowheadResource {
     @Path("notify")
     public Response receiveEvent(Event event) {
         LidarReadout readout = ArrowheadConverter.json().fromString(event.getPayload(), LidarReadout.class);
-        HMIConsumer.window.callback(readout.data);
         return Response.ok().build();
     }
 }
