+++ lidar-common/src/main/java/eu/arrowhead/lidar/common/hmi/HMITheme.java
@@ -1,4 +1,4 @@
-package eu.arrowhead.lidar.consumer;
+package eu.arrowhead.lidar.common.hmi;
 
 import java.awt.*;
